+++ src/App.js
@@ -2,8 +2,8 @@ import React, {useState, useMemo} from "react";
 import './App.less';
 import {Layout} from 'antd';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import HeaderComponent from "./components/HeaderComponent";
-import Club from "./components/club/Club";
+import HeaderComponent from "./components/header/HeaderComponent";
+import ClubComponent from "./components/club/ClubComponent";
 import FooterComponent from "./components/FooterComponent";
 import ProjectComponent from "./components/ProjectComponent";
 import {SearchContext} from "./context/SearchContext";
@@ -27,9 +27,9 @@ function App() {
                     <HeaderComponent/>
                     <Content className="content" style={{padding: '20px 50px', 'min-height': '83vh'}}>
                         <Switch>
-                            <Route path="/clubs" component={Club}/>
+                            <Route path="/clubs" component={ClubComponent}/>
                             <Route path="/projects" component={ProjectComponent}/>
-                            <Route path="/" component={Club}/>
+                            <Route path="/" component={ClubComponent}/>
                         </Switch>
                     </Content>
                 </SearchContext.Provider>
